requirements.txt
asgiref==3.7.2
astroid==2.4.2
Django==4.0
django-filter==23.5
djangorestframework==3.14.0
dnspython==2.4.2
email-validator==2.1.0.post1
Faker==20.1.0
faker-vehicle==0.2.0
idna==3.6
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-client==0.0.1
mysqlclient==2.2.0
pylint==2.5.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typing_extensions==4.9.0
validate-docbr==1.10.0
wrapt==1.12.1
